move examples -> guides
