--- docs/source/concepts/index.rst
@@ -1,8 +1,0 @@
-Concepts
-========
-
-.. toctree::
-   :maxdepth: 1
-
-   sampler
-
