--- docs/source/tutorials/README.rst
@@ -1,4 +1,0 @@
-This is my gallery
-==================
-
-Below is a gallery of examples
